+++ src/pycentauri/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
     "discover",
 ]
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"